+++ src/arg.ts
@@ -1,11 +1,11 @@
 import {text} from '../util/string';
 import {generateForSelf, generateForChild} from '../util/lint';
 import {BoundingRect} from '../lib/rect';
-import Parser from '../index';
 import {Token} from './index';
 import {AtomToken} from './atom';
 import {HiddenToken} from './hidden';
 import type {
+	Config,
 	LintError,
 	AST,
 } from '../base';
@@ -14,6 +14,7 @@ import type {
 
 import {classes} from '../util/constants';
 import {Shadow} from '../util/debug';
+import Parser from '../index';
 
 /* NOT FOR BROWSER END */
 
@@ -55,7 +56,7 @@ export abstract class ArgToken extends Token {
 	/* NOT FOR BROWSER END */
 
 	/** @param parts 以'|'分隔的各部分 */
-	constructor(parts: readonly string[], config = Parser.getConfig(), accum: Token[] = []) {
+	constructor(parts: readonly string[], config: Config, accum: Token[] = []) {
 		super(undefined, config, accum, {
 			AtomToken: 0, Token: 1, HiddenToken: '2:',
 		});